Set epochs in plot to int

+++ src/rectangle/utils/train.py
@@ -206,15 +206,15 @@ class Trainer(nn.Module):
 
         if self.ensemble:
             plt.figure(figsize=(8,6))
-            plt.plot(np.linspace(0,self.nb_epochs,self.nb_epochs), \
+            plt.plot(np.linspace(0,self.nb_epochs,self.nb_epochs, dtype=int), \
                 np.mean(loss_log_ensemble, axis=0))
-            plt.fill_between(np.linspace(0,self.nb_epochs,self.nb_epochs), \
+            plt.fill_between(np.linspace(0,self.nb_epochs,self.nb_epochs, dtype=int), \
                 np.mean(loss_log_ensemble, axis=0)+np.std(loss_log_ensemble, axis=0),\
                 np.mean(loss_log_ensemble, axis=0)-np.std(loss_log_ensemble, axis=0),\
                 alpha=0.3)
             plt.plot(np.linspace(0,self.nb_epochs,\
-                self.nb_epochs//self.val_interval), np.mean(dice_log_ensemble, axis=0))
-            plt.fill_between(np.linspace(0,self.nb_epochs,int(self.nb_epochs//self.val_interval)), \
+                int(self.nb_epochs//self.val_interval), dtype=int), np.mean(dice_log_ensemble, axis=0))
+            plt.fill_between(np.linspace(0,self.nb_epochs,int(self.nb_epochs//self.val_interval), dtype=int), \
                 np.mean(dice_log_ensemble, axis=0)+np.std(dice_log_ensemble, axis=0),\
                 np.mean(dice_log_ensemble, axis=0)-np.std(dice_log_ensemble, axis=0),\
                 alpha=0.3)
@@ -235,9 +235,9 @@ class Trainer(nn.Module):
                 dice_log_ensemble, delimiter=',')
         else:
             plt.figure(figsize=(8,6))
-            plt.plot(np.linspace(0,self.nb_epochs,self.nb_epochs), loss_log)
+            plt.plot(np.linspace(0,self.nb_epochs,self.nb_epochs, dtype=int), loss_log)
             plt.plot(np.linspace(0,self.nb_epochs,\
-                int(self.nb_epochs//self.val_interval)), dice_log)
+                int(self.nb_epochs//self.val_interval), dtype=int), dice_log)
             plt.xlabel('Epoch #')
             plt.legend(['Train Loss', 'Validation Dice'])
             path_ = path.join(self.outdir,\
